+++ click_extra/platforms.py
@@ -33,6 +33,8 @@ import sys
 from dataclasses import dataclass, field
 from typing import Iterable
 
+from itertools import combinations
+
 from . import cache
 
 """ Below is the collection of heuristics used to identify each platform.
@@ -489,6 +491,76 @@ ALL_GROUPS: frozenset[Group] = frozenset(NON_OVERLAPPING_GROUPS | EXTRA_GROUPS)
 
 ALL_OS_LABELS: frozenset[str] = frozenset(p.name for p in ALL_PLATFORMS.platforms)
 """Sets of all recognized labels."""
+
+
+def reduce(items: Iterable[Group | Platform]) -> set[Group | Platform]:
+    """Reduce a collection of ``Group`` and ``Platform`` to a minimal set.
+
+    Returns a deduplicated set of ``Group`` and ``Platform`` that covers the same exact
+    platforms as the original input, but group as much platforms as possible, to reduce
+    the number of items.
+
+    .. hint::
+        Maybe this could be solved with some `Euler diagram
+        <https://en.wikipedia.org/wiki/Euler_diagram>`_ algorithms, like those
+        implemented in `eule <https://github.com/trouchet/eule>`_.
+    """
+    # Collect all platforms.
+    platforms: set[Platform] = set()
+    for item in items:
+        if isinstance(item, Group):
+            platforms.update(item.platforms)
+        else:
+            platforms.add(item)
+
+    # List any group matching the platforms.
+    valid_groups: set[Group] = set()
+    for group in ALL_GROUPS:
+        if group.issubset(platforms):
+            valid_groups.add(group)
+
+    # Test all combination of groups to find the smallest set of groups + platforms.
+    min_items: int = 0
+    results: list[set[Group | Platform]] = []
+    # Serialize group sets for deterministic lookups. Sort them by platform count.
+    groups = tuple(sorted(valid_groups, key=len, reverse=True))
+    for subset_size in range(1, len(groups) + 1):
+
+        # If we already have a solution that involves less items than the current subset of
+        # groups we're going to evaluates, there is no point in continuing.
+        if min_items and subset_size > min_items:
+            break
+
+        for group_subset in combinations(groups, subset_size):
+            # If any group overlaps another, there is no point in exploring this subset.
+            if not all(g[0].isdisjoint(g[1]) for g in combinations(group_subset, 2)):
+                continue
+
+            # Remove all platforms covered by the groups.
+            ungrouped_platforms = platforms.copy()
+            for group in group_subset:
+                ungrouped_platforms.difference_update(group.platforms)
+
+            # Merge the groups and the remaining platforms.
+            reduction = ungrouped_platforms.union(group_subset)
+            reduction_size = len(reduction)
+
+            # Reset the results if we have a new solution that is better than the previous ones.
+            if not results or reduction_size < min_items:
+                results = [reduction]
+                min_items = reduction_size
+            # If the solution is as good as the previous one, add it to the results.
+            elif reduction_size == min_items:
+                results.append(reduction)
+
+    if len(results) > 1:
+        raise RuntimeError(f"Multiple solutions found: {results}")
+
+    # If no reduceted solution was found, return the original platforms.
+    if not results:
+        return platforms
+
+    return results.pop()
 
 
 @cache